+++ src/main/java/com/verdantartifice/primalmagic/common/containers/ConcocterContainer.java
@@ -93,26 +93,22 @@ public class ConcocterContainer extends AbstractContainerMenu {
                 }
                 slot.onQuickCraft(slotStack, stack);
             } else if (index >= 11 && index < 38) {
-                // If transferring from the player's backpack, put wands in the wand slot and everything else into the input slots
+                // If transferring from the player's backpack, put wands in the wand slot and everything else into the input slots or hotbar, in that order
                 if (this.wandSlot.mayPlace(slotStack)) {
                     if (!this.moveItemStackTo(slotStack, 9, 10, false)) {
                         return ItemStack.EMPTY;
                     }
-                } else {
-                    if (!this.moveItemStackTo(slotStack, 0, 9, false)) {
-                        return ItemStack.EMPTY;
-                    }
+                } else if (!this.moveItemStackTo(slotStack, 0, 9, false) && !this.moveItemStackTo(slotStack, 38, 47, false)) {
+                    return ItemStack.EMPTY;
                 }
             } else if (index >= 38 && index < 47) {
-                // If transferring from the player's hotbar, put wands in the wand slot and everything else into the input slots
+                // If transferring from the player's hotbar, put wands in the wand slot and everything else into the input slots or backpack, in that order
                 if (this.wandSlot.mayPlace(slotStack)) {
                     if (!this.moveItemStackTo(slotStack, 9, 10, false)) {
                         return ItemStack.EMPTY;
                     }
-                } else {
-                    if (!this.moveItemStackTo(slotStack, 0, 9, false)) {
-                        return ItemStack.EMPTY;
-                    }
+                } else if (!this.moveItemStackTo(slotStack, 0, 9, false) && !this.moveItemStackTo(slotStack, 11, 38, false)) {
+                    return ItemStack.EMPTY;
                 }
             } else if (!this.moveItemStackTo(slotStack, 11, 47, false)) {
                 // Move all other transfers into the backpack or hotbar