+++ changelog.md
@@ -13,6 +13,7 @@ v0.0.12-alpha
 * Disable hopper interaction with Essence Furnaces; they're too primitive!
 * Fix hopper interaction with various blocks
 * Rebalance calcinator cook times
+* Fix shift-click behavior for Arcane Workbench and Concocter
 
 v0.0.11-alpha
 -------------